Minor refactors & improvements
