+++ src/components/atoms/ProgressBar/ProgressBar.tsx
@@ -1,27 +1,27 @@
 import React from 'react';
 import { Box, VStack } from '@react-native-material/core';
 import { Text } from 'react-native';
-import { colors } from '../../../design/tokens/colors';
+import { colors } from '../../../design/tokens';
 import { styles } from './ProgressBar.style';
 
 interface ProgressBarProps {
-  completed: number;
+  noteCount: number;
   size?: number;
 }
 
-export const ProgressBar = ({ completed, size = 1 }: ProgressBarProps) => {
+export const ProgressBar = ({ noteCount, size = 1 }: ProgressBarProps) => {
   return (
     <VStack items="center">
       <Box h={10} w="100%" bg={colors.black} style={styles.br}>
         <Box
           h={10}
-          w={`${(completed / size) * 100}%`}
+          w={`${(noteCount / size) * 100}%`}
           bg={colors.darkGreen}
           style={styles.br}
         />
       </Box>
       <Box m={2}>
-        <Text style={styles.text}>{`${completed}/${size}`}</Text>
+        <Text style={styles.text}>{`${noteCount}/${size}`}</Text>
       </Box>
     </VStack>
   );
